+++ packages/client/src/components/purchase/ManaPoolBulkButton.tsx
@@ -11,14 +11,15 @@ import Text from '../base/Text';
 
 interface MassBuyButtonProps {
   cards: Card[];
+  block?: boolean;
 }
 
-const ManaPoolBulkButton: React.FC<MassBuyButtonProps> = ({ cards }) => {
+const ManaPoolBulkButton: React.FC<MassBuyButtonProps> = ({ cards, block = true }) => {
   const price = useMemo(() => cards.reduce((acc, card) => acc + (cardPriceManaPool(card) ?? 0), 0), [cards]);
 
   return (
-    <Button type="link" block outline color="accent" href={getBulkManaPoolLink(cards)}>
-      <Flexbox direction="row" justify="between" className="w-full">
+    <Button type="link" block={block} outline color="accent" href={getBulkManaPoolLink(cards)}>
+      <Flexbox direction="row" justify="between" gap="2" className="w-full">
         <Text semibold>Mana Pool</Text>
         {price > 0 && <Text semibold>{`$${price.toFixed(2)}`}</Text>}
       </Flexbox>